+++ dialer_test.go
@@ -24,9 +24,9 @@ import (
 	"testing"
 	"time"
 
-	"cloud.google.com/go/cloudsqlconn/errtype"
-	"cloud.google.com/go/cloudsqlconn/internal/cloudsql"
-	"cloud.google.com/go/cloudsqlconn/internal/mock"
+	"github.com/funayman/cloud-sql-go-connector/errtype"
+	"github.com/funayman/cloud-sql-go-connector/internal/cloudsql"
+	"github.com/funayman/cloud-sql-go-connector/internal/mock"
 	"golang.org/x/oauth2"
 )